Revert "1. product list 페이지 완료 2. 입력/수정페이지 jsp 수정중(지역 js 개발 필요, 수정 update, insert 구분 개발 필요)-박미선"
This reverts commit b78e5cb400e2bed58c44b58779b120dedc1365a6.

+++ GLtest01/src/main/java/com/gl/master/model/product/ProductVo.java
@@ -17,8 +17,8 @@ public class ProductVo {
 	private Date startd;
 	private String img;
 	private String thumb;
-	private double discount;
-	private double point;
+	private int discount;
+	private int point;
 	private int prio;
 	private int event;
 	public ProductVo() {
@@ -132,16 +132,16 @@ public class ProductVo {
 	public void setThumb(String thumb) {
 		this.thumb = thumb;
 	}
-	public double getDiscount() {
+	public int getDiscount() {
 		return discount;
 	}
-	public void setDiscount(double discount) {
+	public void setDiscount(int discount) {
 		this.discount = discount;
 	}
-	public double getPoint() {
+	public int getPoint() {
 		return point;
 	}
-	public void setPoint(double point) {
+	public void setPoint(int point) {
 		this.point = point;
 	}
 	public int getPrio() {
